+++ calliope/__main__.py
@@ -1,0 +1,3 @@
+from audio_input import get_audio_input
+
+get_audio_input()
