+++ pyproject.toml
@@ -1,0 +1,18 @@
+[tool.poetry]
+name = "calliope"
+version = "0.1.0"
+description = ""
+authors = ["David Meredith <daviddmeredith@gmail.com>"]
+license = "ISC"
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.10"
+scipy = "^1.14.1"
+numpy = "^2.1.3"
+pyaudio = "^0.2.14"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
